# chat-backend/docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "TAI后端服务",
    "title": "TAI Backend API",
    "version": "1.0.0"
  },
  "host": "localhost:9090",
  "basePath": "/api/v1",
  "paths": {
    "/chat/conversations": {
      "get": {
        "description": "分页获取用户的对话列表，支持页码和每页数量参数",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Chat"
        ],
        "summary": "获取对话列表",
        "operationId": "getConversations",
        "parameters": [
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 20,
            "description": "每页数量",
            "name": "page_size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "ConversationListResponse",
            "schema": {
              "$ref": "#/definitions/ConversationListResponse"
            }
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      },
      "post": {
        "description": "根据配置创建一个新的对话",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Chat"
        ],
        "summary": "创建对话",
        "operationId": "createConversation",
        "parameters": [
          {
            "description": "对话设置",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "对话设置",
              "type": "object",
              "$ref": "#/definitions/ConversationSettings"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation",
            "schema": {
              "$ref": "#/definitions/Conversation"
            }
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/chat/conversations/{id}": {
      "delete": {
        "description": "根据对话ID删除指定的对话记录和相关数据",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Chat"
        ],
        "summary": "删除对话",
        "operationId": "deleteConversation",
        "parameters": [
          {
            "type": "integer",
            "description": "对话ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageOnlyResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/chat/conversations/{id}/history": {
      "get": {
        "description": "根据对话ID获取该对话的完整消息历史记录",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Chat"
        ],
        "summary": "获取对话历史",
        "operationId": "getConversationHistory",
        "parameters": [
          {
            "type": "integer",
            "description": "对话ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ConversationHistoryResponse",
            "schema": {
              "$ref": "#/definitions/ConversationHistoryResponse"
            }
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/chat/conversations/{id}/settings": {
      "get": {
        "description": "根据对话ID获取该对话的详细设置信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Chat"
        ],
        "summary": "获取对话设置",
        "operationId": "getConversationSettings",
        "parameters": [
          {
            "type": "integer",
            "description": "对话ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ConversationSettings",
            "schema": {
              "$ref": "#/definitions/ConversationSettings"
            }
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      },
      "put": {
        "description": "根据对话ID更新该对话的设置信息，如模型配置、参数等",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Chat"
        ],
        "summary": "更新对话设置",
        "operationId": "updateConversationSettings",
        "parameters": [
          {
            "type": "integer",
            "description": "对话ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "对话设置",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "对话设置",
              "type": "object",
              "$ref": "#/definitions/ConversationSettings"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageWithDataResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/chat/messages": {
      "post": {
        "description": "向指定对话发送消息并以SSE流式方式返回AI响应",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/event-stream"
        ],
        "tags": [
          "Chat"
        ],
        "summary": "发送消息",
        "operationId": "sendMessage",
        "parameters": [
          {
            "description": "聊天请求",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "聊天请求",
              "type": "object",
              "$ref": "#/definitions/ChatRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/knowledge/bases": {
      "get": {
        "description": "获取系统中所有已创建的知识库列表，包括知识库基本信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Knowledge"
        ],
        "summary": "获取知识库列表",
        "operationId": "listKnowledgeBases",
        "responses": {
          "200": {
            "$ref": "#/responses/models.KnowledgeBase"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      },
      "post": {
        "description": "创建一个新的知识库，用于存储和管理相关文档",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Knowledge"
        ],
        "summary": "创建知识库",
        "operationId": "createKnowledgeBase",
        "parameters": [
          {
            "description": "知识库信息",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "知识库信息",
              "type": "object",
              "$ref": "#/definitions/KnowledgeBaseCreateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageWithDataResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/knowledge/bases/{id}": {
      "put": {
        "description": "更新指定知识库的名称、描述等基本信息",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Knowledge"
        ],
        "summary": "更新知识库",
        "operationId": "updateKnowledgeBase",
        "parameters": [
          {
            "type": "integer",
            "description": "知识库ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "更新信息",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "更新信息",
              "type": "object",
              "$ref": "#/definitions/UpdateKnowledgeBaseRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageWithDataResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      },
      "delete": {
        "description": "根据知识库ID删除指定的知识库及其所有相关文件",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Knowledge"
        ],
        "summary": "删除知识库",
        "operationId": "deleteKnowledgeBase",
        "parameters": [
          {
            "type": "integer",
            "description": "知识库ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageOnlyResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/knowledge/bases/{id}/files": {
      "get": {
        "description": "获取指定知识库中所有已上传文件的列表和状态信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Knowledge"
        ],
        "summary": "获取知识库文件",
        "operationId": "getKnowledgeBaseFiles",
        "parameters": [
          {
            "type": "integer",
            "description": "知识库ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/models.KnowledgeFile"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      },
      "post": {
        "description": "支持两种上传方式：\n文件流上传：使用 multipart/form-data 格式，字段名为 'file',多个文件使用多个 'file' 字段\n文件路径上传：使用 application/json 格式，body 中包含 file_paths 数组",
        "consumes": [
          "multipart/form-data",
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Knowledge"
        ],
        "summary": "上传文件到知识库",
        "operationId": "uploadFile",
        "parameters": [
          {
            "type": "integer",
            "description": "知识库ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "object",
            "description": "上传文件（支持单个或多个文件）",
            "name": "file"
          },
          {
            "description": "单文件路径或多文件路径列表（用于路径上传）",
            "name": "body",
            "in": "body",
            "schema": {
              "description": "单文件路径或多文件路径列表（用于路径上传）",
              "type": "object",
              "$ref": "#/definitions/BatchUploadFilesRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "BatchUploadResponse",
            "schema": {
              "$ref": "#/definitions/BatchUploadResponse"
            }
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/knowledge/files/{file_id}": {
      "delete": {
        "description": "从知识库中删除指定的文件及其向量数据",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Knowledge"
        ],
        "summary": "删除知识库文件",
        "operationId": "deleteFile",
        "parameters": [
          {
            "type": "integer",
            "description": "文件ID",
            "name": "file_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageOnlyResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/knowledge/files/{file_id}/toggle": {
      "put": {
        "description": "启用或禁用知识库中的指定文件，控制其是否参与问答检索",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Knowledge"
        ],
        "summary": "切换文件启用状态",
        "operationId": "toggleFileEnable",
        "parameters": [
          {
            "type": "integer",
            "description": "文件ID",
            "name": "file_id",
            "in": "path",
            "required": true
          },
          {
            "description": "状态切换",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "状态切换",
              "type": "object",
              "$ref": "#/definitions/FileToggleEnableRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageOnlyResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/models": {
      "post": {
        "description": "保存或更新模型的配置信息，包括API密钥、端点等设置",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Models"
        ],
        "summary": "保存模型配置",
        "operationId": "saveModel",
        "parameters": [
          {
            "description": "模型信息",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "模型信息",
              "type": "object",
              "$ref": "#/definitions/ModelSaveRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ModelID",
            "schema": {
              "$ref": "#/definitions/ModelID"
            }
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/models/available/all": {
      "get": {
        "description": "获取当前系统中所有可用的模型列表，包括聊天模型和向量模型",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Models"
        ],
        "summary": "获取所有可用模型",
        "operationId": "listAvailableAllModels",
        "responses": {
          "200": {
            "$ref": "#/responses/models.ModelInfo"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/models/available/chat": {
      "get": {
        "description": "获取当前系统中可用的聊天模型列表，仅包含已配置且可用的聊天模型",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Models"
        ],
        "summary": "获取可用聊天模型",
        "operationId": "listAvailableChatModels",
        "responses": {
          "200": {
            "$ref": "#/responses/models.ModelInfo"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/models/available/vector": {
      "get": {
        "description": "获取当前系统中可用的向量模型列表，用于知识库文档向量化",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Models"
        ],
        "summary": "获取可用向量模型",
        "operationId": "listAvailableVectorModels",
        "responses": {
          "200": {
            "$ref": "#/responses/models.ModelInfo"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/models/supported/chat": {
      "get": {
        "description": "获取系统支持的所有聊天模型列表，包括模型名称、类型等信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Models"
        ],
        "summary": "获取支持的聊天模型",
        "operationId": "listSupportedChatModels",
        "responses": {
          "200": {
            "$ref": "#/responses/models.SupportedChatModel"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/models/supported/vector": {
      "get": {
        "description": "获取系统支持的所有向量模型列表，用于知识库向量化处理",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Models"
        ],
        "summary": "获取支持的向量模型",
        "operationId": "listSupportedVectorModels",
        "responses": {
          "200": {
            "$ref": "#/responses/models.SupportedVectorModel"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/models/{id}": {
      "delete": {
        "description": "根据模型ID删除指定的模型配置信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Models"
        ],
        "summary": "删除模型配置",
        "operationId": "deleteModel",
        "parameters": [
          {
            "type": "integer",
            "description": "模型ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageOnlyResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/models/{id}/status": {
      "put": {
        "description": "启用或禁用指定的模型，控制模型是否可用于对话",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Models"
        ],
        "summary": "设置模型状态",
        "operationId": "setModelStatus",
        "parameters": [
          {
            "type": "integer",
            "description": "模型ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "状态设置",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "状态设置",
              "type": "object",
              "$ref": "#/definitions/ModelStatusEnableRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageOnlyResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/settings/defaults": {
      "get": {
        "description": "获取系统当前的默认配置信息，包括模型设置、聊天参数等",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Settings"
        ],
        "summary": "获取默认设置",
        "operationId": "getDefaultSettings",
        "responses": {
          "200": {
            "description": "DefaultSettings",
            "schema": {
              "$ref": "#/definitions/DefaultSettings"
            }
          }
        }
      },
      "put": {
        "description": "更新系统的默认配置信息，新的设置将应用于后续创建的对话",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Settings"
        ],
        "summary": "更新默认设置",
        "operationId": "updateDefaultSettings",
        "parameters": [
          {
            "description": "默认设置",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "默认设置",
              "type": "object",
              "$ref": "#/definitions/DefaultSettings"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "DefaultSettings",
            "schema": {
              "$ref": "#/definitions/DefaultSettings"
            }
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/settings/defaults/reset": {
      "post": {
        "description": "将系统配置重置为出厂默认值，恢复所有设置到初始状态",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Settings"
        ],
        "summary": "重置默认设置",
        "operationId": "resetDefaultSettings",
        "responses": {
          "200": {
            "description": "DefaultSettings",
            "schema": {
              "$ref": "#/definitions/DefaultSettings"
            }
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BatchUploadFilesRequest": {
      "description": "BatchUploadFilesRequest 批量上传文件请求",
      "type": "object",
      "required": [
        "file_paths"
      ],
      "properties": {
        "file_paths": {
          "description": "上传文件路径列表（本地文件系统路径）",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "FilePaths"
        }
      },
      "x-go-package": "chat-backend/models"
    },
...
    "ConversationHistoryResponse": {
      "description": "ConversationHistoryResponse 对话历史响应",
      "type": "object",
      "required": [
        "conversation_id",
        "messages",
        "total"
      ],
      "properties": {
        "conversation_id": {
          "description": "对话ID",
          "type": "string",
          "x-go-name": "ConversationID"
        },
        "messages": {
          "description": "消息列表",
          "type": "array",
          "items": {
            "$ref": "#/definitions/MessageRecord"
          },
          "x-go-name": "Messages"
        },
        "total": {
          "description": "消息总数",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Total"
        }
      },
      "x-go-package": "chat-backend/models"
    },
...
    "ConversationListResponse": {
      "description": "ConversationListResponse 对话列表响应",
      "type": "object",
      "required": [
        "conversations",
        "total",
        "page",
        "page_size"
      ],
      "properties": {
        "conversations": {
          "description": "对话列表",
          "type": "array",
          "items": {
            "$ref": "#/definitions/Conversation"
          },
          "x-go-name": "Conversations"
        },
        "page": {
          "description": "当前页码",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Page"
        },
        "page_size": {
          "description": "每页数量",
          "type": "integer",
          "format": "int64",
          "x-go-name": "PageSize"
        },
        "total": {
          "description": "总数",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Total"
        }
      },
      "x-go-package": "chat-backend/models"
    },
    "ConversationSettings": {
      "description": "ConversationSettings 对话设置",
      "type": "object",
      "required": [
        "name",
        "desc",
        "model_id",
        "temperature",
        "top_p",
        "presence_penalty",
        "frequency_penalty",
        "responseType",
        "stream",
        "knowledge_base_ids",
        "contextLimit"
      ],
      "properties": {
        "contextLimit": {
          "description": "上下文限制（消息数量）",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ContextLimit"
        },
        "desc": {
          "description": "对话描述",
          "type": "string",
          "x-go-name": "Desc"
        },
        "frequency_penalty": {
          "description": "重复控制",
          "type": "number",
          "format": "double",
          "x-go-name": "FrequencyPenalty"
        },
        "knowledge_base_ids": {
          "description": "知识库ID列表",
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int64"
          },
          "x-go-name": "KnowledgeBaseIDs"
        },
        "model_id": {
          "description": "模型ID（数字）",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ModelID"
        },
        "name": {
          "description": "对话名称",
          "type": "string",
          "x-go-name": "Name"
        },
        "presence_penalty": {
          "description": "词汇控制",
          "type": "number",
          "format": "double",
          "x-go-name": "PresencePenalty"
        },
        "responseType": {
          "description": "响应类型  text/json",
          "type": "string",
          "x-go-name": "ResponseType"
        },
        "stream": {
          "description": "对话输出模式 true=流式输出, false=非流式输出",
          "type": "boolean",
          "x-go-name": "Stream"
        },
        "temperature": {
          "description": "多样性",
          "type": "number",
          "format": "double",
          "x-go-name": "Temperature"
        },
        "top_p": {
          "description": "采样范围",
          "type": "number",
          "format": "double",
          "x-go-name": "TopP"
        }
      },
      "x-go-package": "chat-backend/models"
    },
...
    "ResponseBody": {
      "description": "ResponseBody 统一响应体结构",
      "type": "object",
      "required": [
        "is_login",
        "code",
        "msg",
        "data",
        "time",
        "code_num",
        "code_msg"
      ],
      "properties": {
        "code": {
          "description": "请求是否成功",
          "type": "boolean",
          "x-go-name": "Code"
        },
        "code_msg": {
          "description": "错误代码消息",
          "type": "string",
          "x-go-name": "CodeMsg"
        },
        "code_num": {
          "description": "数字错误代码",
          "type": "integer",
          "format": "int64",
          "x-go-name": "CodeNum"
        },
        "data": {
          "description": "响应数据",
          "x-go-name": "Data"
        },
        "is_login": {
          "description": "登录状态",
          "type": "boolean",
          "x-go-name": "IsLogin"
        },
        "msg": {
          "description": "响应消息",
          "type": "string",
          "x-go-name": "Message"
        },
        "time": {
          "description": "持续时间（秒）",
          "type": "number",
          "format": "double",
          "x-go-name": "Time"
        }
      },
      "x-go-package": "chat-backend/utils"
    },
...
  },
  "responses": {
...
    "MessageOnlyResponse": {
      "description": "MessageOnlyResponse 仅消息成功响应",
      "schema": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "description": "响应消息",
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
    "MessageWithDataResponse": {
      "description": "MessageWithDataResponse 消息和数据成功响应",
      "schema": {
        "type": "object",
        "required": [
          "message",
          "data"
        ],
        "properties": {
          "data": {
            "description": "响应数据",
            "x-go-name": "Data"
          },
          "message": {
            "description": "响应消息",
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
...
  }
}